Update db schema

+++ server/migrations/2020-03-07_23-39-24_AddAlbum2Image.sql
@@ -1,6 +1,6 @@
 -- Add migration script here
 CREATE TABLE album2image (
-    album_id INTEGER NOT NULL REFERENCES album(id),
-    image_id INTEGER NOT NULL REFERENCES image(id),
+    album_id BIGINT NOT NULL REFERENCES album(id),
+    image_id BIGINT NOT NULL REFERENCES image(id),
     PRIMARY KEY (album_id,image_id)
 );
